Finished re-gigging around location and grouping of tests. #47 - changed default behaviour, so that fast-start isn't used by default - it now has to be explicitly enabled through setting. Added long parameter support to bean utils.

+++ src/main/java/org/littlegrid/ClusterMemberGroup.java
@@ -501,7 +501,7 @@ public interface ClusterMemberGroup {
          * @param joinTimeoutMilliseconds Join timeout milliseconds.
          * @return builder.
          */
-        Builder setFastStartJoinTimeoutMilliseconds(int joinTimeoutMilliseconds);
+        Builder setFastStartJoinTimeoutMilliseconds(long joinTimeoutMilliseconds);
     }
 
     /**
